feat: Styled descendant elements
Add multiple versions of descendant combinators to style the elements as
required by this task. Each rule works to style the elements but only 1
of the 4 is actually needed (hence the last 3 are commented out).

All 4 have been included as practice for myself and a reminder should I
need to come back and revise this topic.